+++ src/coq_elpi_utils.ml
@@ -210,7 +210,7 @@ let detype_closed_glob env sigma closure =
   fix_detype gbody
 
 type qualified_name = string list
-let compare_qualified_name = Pervasives.compare
+let compare_qualified_name = Stdlib.compare
 let pr_qualified_name = Pp.prlist_with_sep (fun () -> Pp.str".") Pp.str
 let show_qualified_name = String.concat "."
 let pp_qualified_name fmt l = Format.fprintf fmt "%s" (String.concat "." l)