+++ Changelog.md
@@ -14,6 +14,8 @@ Requires Elpi 1.13.6 and Coq 8.13.
 - New `coq.hints.add-mode` like the `Hint Mode` vernacular
 - New `coq.hints.modes`
 - New `coq.TC.declare-class`
+- Deprecate `coq.env.const-opaque?` -> `coq.env.opaque?`
+- Deprecate `coq.env.const-primitive?` -> `coq.env.primitive?`
 
 ## [1.10.3] - 18-06-2021